+++ app/api/v1/api.py
@@ -6,6 +6,6 @@ api_router = APIRouter()
 
 # Include endpoint routers
 api_router.include_router(status.router, prefix="/app_status", tags=["Status"])
-api_router.include_router(parser.router, prefix="/parse_message", tags=["Parser"])
+api_router.include_router(parser.router, prefix="/parser", tags=["Parser"])
 api_router.include_router(rules.router, prefix="/rules", tags=["Rules"])